mpv: converge from morpheus, ravu-lite for the iGPU
Takes morpheus's structure and updated shaders, but keeps luma upscale on
ravu-lite + plain ewa_lanczos - FSRCNNX and ewa_lanczos4sharpest will drop
frames on UHD 620.

Also fixes a live bug: the old config pointed at ravu-lite-r4.hook, which
does not exist. The file is ravu-lite-ar-r4.hook.

+++ .config/mpv/shaders/SSimDownscaler.glsl
@@ -1,7 +1,3 @@
-// Revised 09/25/24
-//
-// https://gist.github.com/igv/36508af3ffc84410fe39761d6969be10
-//
 // This library is free software; you can redistribute it and/or
 // modify it under the terms of the GNU Lesser General Public
 // License as published by the Free Software Foundation; either
@@ -16,8 +12,8 @@
 // License along with this library.
 
 //!HOOK POSTKERNEL
-//!BIND PREKERNEL
 //!BIND HOOKED
+//!BIND PREKERNEL
 //!SAVE L2
 //!WIDTH NATIVE_CROPPED.w
 //!WHEN NATIVE_CROPPED.h POSTKERNEL.h >
@@ -57,8 +53,8 @@ vec4 hook() {
 }
 
 //!HOOK POSTKERNEL
-//!BIND L2
 //!BIND HOOKED
+//!BIND L2
 //!SAVE L2
 //!WHEN NATIVE_CROPPED.w POSTKERNEL.w >
 //!COMPONENTS 3
@@ -155,7 +151,7 @@ vec4 hook() {
 
     float Sl = Luma(max(avg[1] - avg[0] * avg[0], 0.));
     float Sh = Luma(max(avg[2] - avg[0] * avg[0], 0.));
-    return vec4(avg[0], mix(sqrt((Sh + sigma_nsq) / (Sl + sigma_nsq)) * (1. + oversharp), clamp(Sh / Sl, 0., 1.), float(Sl > Sh)));
+    return vec4(avg[0], mix(sqrt((Sh + sigma_nsq) / (Sl + sigma_nsq)) * (1. + oversharp), clamp(Sh / Sl, 0., 1.), int(Sl > Sh)));
 }
 
 //!HOOK POSTKERNEL
